final-qa-project/code/requirements.txt
pytest
pytest-xdist
requests
selenium
webdriver-manager
SQLAlchemy
Faker
Flask
PyMySQL
allure-pytest
